[Rust][refactor](sys, sdk): Refactored code.

+++ bindings/rust/wasmedge-sdk/examples/create_import_object.rs
@@ -12,7 +12,7 @@ use wasmedge_sys::types::WasmValue;
 #[cfg_attr(test, test)]
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     // a native function to be imported as host function
-    fn real_add(inputs: Vec<WasmValue>) -> std::result::Result<Vec<WasmValue>, u8> {
+    fn real_add(inputs: Vec<WasmValue>) -> std::result::Result<Vec<WasmValue>, u32> {
         if inputs.len() != 2 {
             return Err(1);
         }
