+++ bindings/rust/wasmedge-sdk/src/externals/table.rs
@@ -263,7 +263,7 @@ mod tests {
         }
     }
 
-    fn real_add(inputs: Vec<WasmValue>) -> std::result::Result<Vec<WasmValue>, u8> {
+    fn real_add(inputs: Vec<WasmValue>) -> std::result::Result<Vec<WasmValue>, u32> {
         if inputs.len() != 2 {
             return Err(1);
         }
