+++ bindings/rust/wasmedge-sys/examples/call_hostfunc_chain.rs
@@ -43,13 +43,13 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     vm.load_wasm_from_bytes(&wasm_bytes)?;
     vm.validate()?;
 
-    let host_layer1 = |_: Vec<WasmValue>| -> Result<Vec<WasmValue>, u8> {
+    let host_layer1 = |_: Vec<WasmValue>| -> Result<Vec<WasmValue>, u32> {
         println!("There is layer1!");
         Ok(vec![])
     };
 
     let s = Arc::new(Mutex::new(Wrapper(&vm as *const Vm)));
-    let host_layer2 = move |_: Vec<WasmValue>| -> Result<Vec<WasmValue>, u8> {
+    let host_layer2 = move |_: Vec<WasmValue>| -> Result<Vec<WasmValue>, u32> {
         unsafe {
             (*s.lock().unwrap().0).run_function("layer1", []).unwrap();
         }
