+++ bindings/rust/wasmedge-sys/examples/vm_run_registered_func_async.rs
@@ -6,7 +6,7 @@ use wasmedge_sys::{
 use wasmedge_types::ValType;
 
 // A native function
-fn real_add(inputs: Vec<WasmValue>) -> Result<Vec<WasmValue>, u8> {
+fn real_add(inputs: Vec<WasmValue>) -> Result<Vec<WasmValue>, u32> {
     if inputs.len() != 2 {
         return Err(1);
     }
