+++ bindings/rust/wasmedge-sys/src/instance/table.rs
@@ -444,7 +444,7 @@ mod tests {
         handle.join().unwrap();
     }
 
-    fn real_add(input: Vec<WasmValue>) -> Result<Vec<WasmValue>, u8> {
+    fn real_add(input: Vec<WasmValue>) -> Result<Vec<WasmValue>, u32> {
         println!("Rust: Entering Rust function real_add");
 
         if input.len() != 2 {
