+++ bindings/rust/wasmedge-sys/src/store.rs
@@ -339,7 +339,7 @@ mod tests {
         assert_eq!(return_types, [ValType::I32]);
     }
 
-    fn real_add(inputs: Vec<WasmValue>) -> Result<Vec<WasmValue>, u8> {
+    fn real_add(inputs: Vec<WasmValue>) -> Result<Vec<WasmValue>, u32> {
         if inputs.len() != 2 {
             return Err(1);
         }
