+++ bindings/rust/wasmedge-sys/src/vm.rs
@@ -2528,7 +2528,7 @@ mod tests {
     }
 
     #[cfg(unix)]
-    fn real_add(inputs: Vec<WasmValue>) -> Result<Vec<WasmValue>, u8> {
+    fn real_add(inputs: Vec<WasmValue>) -> Result<Vec<WasmValue>, u32> {
         if inputs.len() != 2 {
             return Err(1);
         }
